Make Droppable and Draggable components sound
Currently `Droppable` and `Draggable` components are treated as `any`
which a makes development less safer.

In this diff I with @istarkov help found a way to cover connected
components with default props without loosing type information.
A few basic flow tests will let to not remove this info accidentally.

As a result unused type property was result from some stories.

Some error suppressions are not necessary so I removed them.